Rename some asset directores so capitalization matches the code.
In order for the program to work on a case-sensitive file system, the
capitalization of the actual path names must match what is used in the
code. Otherwise the asset being referenced will not be found. In this
case it would result in the weather icon disappearing from the menu bar.